refactor: cardNickName 관련 로직 customhook으로 분리
